requirements.txt
streamlit==1.51.0
chromadb==1.3.0
sentence-transformers==5.1.2
openai==1.109.1
python-dotenv==1.2.1
pdfplumber==0.11.7
numpy==1.26.4
pydantic==2.8.2